refactor(storage): use nitrogen library for AHCI and NVMe
Replace custom AHCI and NVMe driver implementations with thin wrappers around the `nitrogen::storage::ahci` and `nitrogen::storage::nvme` modules. This reduces kernel code by ~600 lines and consolidates driver logic into the shared `nitrogen` library.

- AHCI: Removed bare-metal HBA register handling and DMA setup, using nitrogen's safe abstractions.
- NVMe: Removed PCIe BAR mapping, queue allocation, and controller initialization, replaced with nitrogen's NVMe subsystem.
